+++ edbee-app/ui/filetreesidewidget.h
@@ -42,8 +42,12 @@ public slots:
     void clearAllRootPaths();
 
     void reveal( const QString& filename );
+
     void startRenameItem( const QString& filename );
     void startRenameItemByAction();
+
+    void createNewFileAndRename( const QString& pathname );
+    void createNewFileAndRenameByAction();
 
 private: